vertex reader and fixes

+++ Examples/Io/Root/include/ActsExamples/Io/Root/RootParticleWriter.hpp
@@ -95,27 +95,27 @@ class RootParticleWriter final : public WriterT<SimParticleContainer> {
   /// Production process type, i.e. what generated the particle.
   std::vector<std::uint32_t> m_process;
   /// Production position components in mm.
-  std::vector<float> m_vx;
-  std::vector<float> m_vy;
-  std::vector<float> m_vz;
-  std::vector<float> m_vt;
+  std::vector<double> m_vx;
+  std::vector<double> m_vy;
+  std::vector<double> m_vz;
+  std::vector<double> m_vt;
   /// Total momentum in GeV
-  std::vector<float> m_p;
+  std::vector<double> m_p;
   /// Momentum components in GeV.
-  std::vector<float> m_px;
-  std::vector<float> m_py;
-  std::vector<float> m_pz;
+  std::vector<double> m_px;
+  std::vector<double> m_py;
+  std::vector<double> m_pz;
   /// Mass in GeV.
-  std::vector<float> m_m;
+  std::vector<double> m_m;
   /// Charge in e.
-  std::vector<float> m_q;
+  std::vector<double> m_q;
   // Derived kinematic quantities
   /// Direction pseudo-rapidity.
-  std::vector<float> m_eta;
+  std::vector<double> m_eta;
   /// Direction angle in the transverse plane.
-  std::vector<float> m_phi;
+  std::vector<double> m_phi;
   /// Transverse momentum in GeV.
-  std::vector<float> m_pt;
+  std::vector<double> m_pt;
   // Decoded particle identifier; see Barcode definition for details.
   std::vector<std::uint32_t> m_vertexPrimary;
   std::vector<std::uint32_t> m_vertexSecondary;
@@ -125,11 +125,11 @@ class RootParticleWriter final : public WriterT<SimParticleContainer> {
 
   // Optional information depending on input collections.
   /// Total energy loss in GeV.
-  std::vector<float> m_eLoss;
+  std::vector<double> m_eLoss;
   /// Accumulated material
-  std::vector<float> m_pathInX0;
+  std::vector<double> m_pathInX0;
   /// Accumulated material
-  std::vector<float> m_pathInL0;
+  std::vector<double> m_pathInL0;
   /// Number of hits.
   std::vector<std::int32_t> m_numberOfHits;
 };
